requirements.txt
asgiref==3.2.3
Django==3.0.3
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
xlrd==1.2.0
